move exception logger, fix exception handling in threads
