+++ pyacq/core/rpc/log/__init__.py
@@ -8,3 +8,4 @@ from .remote import (get_logger_address, set_logger_address,
                      get_thread_name, set_thread_name,
                      start_log_server, LogSender, LogServer)
 from .handler import RPCLogHandler, log_exceptions
+from .exceptions import log_exceptions
